runtime: Inline `extend_and_hash`
#### Problem

`solana_sha256_hasher::extend_and_hash` requires dynamic allocation
because of the vec creation, which means adding a `std` feature to the
crate.

However, the function is only used in one place, so it would be easier
to remove it and simplify the hasher crate.

#### Summary of changes

Inline the usage of `extend_and_hash` so it can be safely removed from
the hasher crate.