poseidon: Enforce padding in Poseidon hash inputs
light-poseidon 0.4.0 introduced a more rigid hash input check
(Lightprotocol/light-poseidon@92ce32a6513b), where every input needs
to have a length matching the byte length of the prime field modulus.
In case of the `Fr` field, it's 32 bytes.

Previously it was possible to pass smaller byte inputs, which still
have the same value as a scalar, e.g. (little-endian):

```
[0, 0, 0, 0, 0, 0, 0, 1]
[0, 0, 0, 1]
[0, 1]
[1]
```

From now on, the only accepted way of passing a scalar `1` will be:

```
[
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 1
]
```

Upgrade of light-poseidon requires upgrade of ark-bn254 from 0.4.0 to
0.5.0

This behavior is guarded by the new feature flag
`poseidon_enforce_padding`. If the feature flag is disabled, the old
versions of light-poseidon and ark-bn254 are used to ensure the old
behavior.