transaction-status-client-types: Fix deserializer for SDKv2
#### Problem

The InstructionError deserializer within UiTransactionError fails to
deserialize errors from SDKv2 because `BorshIoError` changed from being
a newtype with string to a unit enum variant.

#### Summary of changes

If deserialization fails, check to see if it's an old `BorshIoError`,
and change to the new variation if so.

Also include a test using the old and new error formats.