Remove transaction processing benches
Each of consumer and blockstore_processor have a bench that executes
transactions. These benches pregenerate a bunch of transactions to avoid
noise in the results; however, the pre-generated transactions also make
the bench flaky if the entire array of transactions is executed.

The bench has limited value to us at the moment so rip them out instead
of continuing to maintain them