Add agave-xdp-ebpf
Add our first (but not last!) eBPF program. The program is loaded by
agave-xdp and includes logic to workaround a bug with the i40e network
driver.

The eBPF program is pre-built so people don't need to install nightly +
bpf-linker to build. See xdp-ebpf/README for more details.

See
https://lore.kernel.org/netdev/20251021173200.7908-2-alessandro.d@gmail.com
for a description of the i40e driver bug.