xdp: tx_loop: change MAX_TIMEOUTS to 1
Once upon a time we used to spin without sleeping, and we flushed
batches after 500 empty spins on the receive channel. Then we added
sleeping to save CPU but kept MAX_TIMEOUTS too high.

When load is relatively low - like during banking - this can lead to a
non-full batch to be held for up to 500ms, which can cause slots to be
skipped.

Fix by bringing MAX_TIMEOUTS to 1 which means that during banking we'll
hold shreds for up to 1ms.