build-sbf: Make it work in any directory
#### Problem

Currently, users are forced to download the whole tarball just to get
`cargo-build-sbf`, which adds a lot of unnecessary bloat, considering
people are used to simply running `cargo install ...` to get binaries.

#### Summary of changes

Builds on top of #9345 and #9346 to remove the usage of `--sbf-sdk`
everywhere, which makes cargo-build-sbf only depend on the platform
tools install directory.

This means that the `cargo-build-sbf` binary can live anywhere and Just
Work, so it will work in `cargo install` settings too!