blockstore: Chain merkle root for slot 0 to genesis hash
The function create_new_ledger() is used by solana-genesis to create a
new blockstore with slot 0 full of ticks. The chained merkle root for
the shredder is currently initialized to an array of random bytes. This
means that calling this function two times with the same genesis config
will yield different chained merkle roots.

There is no reason for this to be the case and we can make this
deterministic by seeding the chained merkle root for slot 0 with the
genesis hash