streamer: Use `BytesPacket` instead of `Packet`
The `Packet` struct stores the payload in a byte array. Using that
struct often involves moving and cloning it, which then results in
large memory copies.

Furthermore, the recv loop is allocating each packet buffer separately.
For a batch of 64 packets, it does 64 allocations of ~1500 bytes. And
that's happening on each iteration.

`BytesPacket` is the new type of packet (introduced in #5646) which uses
`bytes::Bytes` and therefore doesn't make an actual copy of the
underlying data. We use it in QUIC servers already.

Use it in the UDP streamer as well, to let the other components of Agave
(gossip, turbine etc.) benefit from the zero-copy capabilities.

Instead of doing multiple allocations in a loop, allocate a huge buffer
for the whole batch outside of the loop. After each iteration, just
`reserve()` the necessary space, which in the best case (when the packet
chunks, that were split off, are not used anymore), it will just re-use
the same chunk of memory.

Change the `recvmmsg` and `sendmmsg` wrappers to work with the `bytes`
types. Do that by providing a wrapper type over `BytesMut`, called
`RecvBuffer`, which allows for effective allocation of a large buffer
for the whole batch and yielding the chunks, which can be passed to
`recvmmsg`.