build-sbf: Inline all scripts
#### Problem

cargo-build-sbf should be usable from `cargo install`, but there are a
lot of assumptions about the executable living in a Solana release. One
of those is a connection to an SBF SDK directory.

Upon a longer investigation, I realized that nothing is really necessary
from the SBF SDK directory, and that it's most useful as a link to the
platform tools.

#### Summary of changes

As the first step of removing SBF SDK from build-sbf, inline all of the
scripts used from the SBF SDK, which are dump.sh, objcopy.sh, and
strip.sh.

I tested that this produces almost exactly the same output -- I removed
the first line from dump.sh, which does: `ls -la "$so" > "$dump_mangled"