nextest: limit sbf tests to 1 concurrent
These tests all might download a tarball of the toolchain and when these
tests are run concurrently each such download steps on each other toes,
leading to pretty confusing failures.

CI profile already limits the execution to 1 parallel test of this
category for what seems to be exactly the same reason.