build-sbf: patch the binaries in downloaded toolchain for NixOS users
This effectively duplicates similar logic from rustc's build process
that gets applied to the downloaded bootstrap compiler, some of which
was written by… myself.

This is a relatively selfish PR, I'll admit, as I'm using NixOS and want
to be able to run the test suite without having to jump through the
hoops of setting up the environment tailored for the tests.

As far as I understand `build-sbf`, it appears to be end-user facing and
therefore should now allow anybody using NixOS to use `build-sbf` with a
similar level of experience as people who use a more traditional
distribution get.